refactor: move rewind online tests to dedicated CI job
Move rewind-feature.test.ts and selective-rewind.test.ts from
tests/online/session/features/ to tests/online/rewind/ so they
run as a separate Daemon Online (rewind) CI job.
