+++ packages/daemon/tests/online/rewind/selective-rewind.test.ts
@@ -17,9 +17,9 @@
 
 import { describe, test, expect, beforeEach, afterEach } from 'bun:test';
 import 'dotenv/config';
-import type { DaemonServerContext } from '../../helpers/daemon-server-helper';
-import { createDaemonServer } from '../../helpers/daemon-server-helper';
-import { sendMessage, waitForIdle } from '../../helpers/daemon-test-helpers';
+import type { DaemonServerContext } from '../helpers/daemon-server-helper';
+import { createDaemonServer } from '../helpers/daemon-server-helper';
+import { sendMessage, waitForIdle } from '../helpers/daemon-test-helpers';
 import type { RewindMode } from '@neokai/shared';
 
 /**
